moved LogLevel env variable parsing to comgr-env.cpp
